Get to sequence encoding...
