+++ zstd/tozstd/tozstd.go
@@ -2,7 +2,7 @@ package tozstd
 
 import "log"
 
-const debug = false
+const debug = true
 
 func println(a ...interface{}) {
 	if debug {
